% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taglasso.R
\name{taglasso}
\alias{taglasso}
\title{tag-lasso estimation of the precision matrix}
\usage{
taglasso(
  X,
  A,
  pendiag = F,
  lambda1,
  lambda2,
  rho = 10^-2,
  it_in = 100,
  it_out = 10,
  it_in_refit = 1000,
  it_out_refit = 20,
  hc = TRUE
)
}
\arguments{
\item{X}{An (\eqn{n}x\eqn{p})-matrix of \eqn{p} variables and \eqn{n} observations}

\item{A}{An (\eqn{p}x\eqn{|T|})- binary matrix incorporating the tree-based aggregation structure}

\item{pendiag}{Logical indicator whether or not to penalize the diagonal in Omega. The default is \code{TRUE} (penalization of the diagonal)}

\item{lambda1}{Aggregation tuning parameter. Use \code{taglasso_cv} to let the program determine this tuning parameter based on K-fold cross-validation}

\item{lambda2}{Sparsity tuning parameter. Use \code{taglasso_cv} to let the program determine this tuning parameter based on K-fold cross-validation}

\item{rho}{Starting value for the LA-ADMM tuning parameter. Default is 10^2; will be locally adjusted via LA-ADMM}

\item{it_in}{Number of inner stages of the LA-ADMM algorithm. Default is 100}

\item{it_out}{Number of outer stages of the LA-ADMM algorithm. Default is 10}

\item{it_in_refit}{Number of inner stages of the LA-ADMM algorithm for re-fitting. Default is 100}

\item{it_out_refit}{Number of outer stages of the LA-ADMM algorithm for re-fitting. Default is 10}

\item{hc}{Logical indicator whether or not hierarchical clustering on the aggregated precision matrix needs to be performed to help visualize the block-structure}
}
\value{
A list with the following components
\item{\code{omega_full}}{Estimated (\eqn{p}x\eqn{p}) precision matrix}
\item{\code{omega_aggregated}}{Estimated (\eqn{K}x\eqn{K}) precision matrix}
\item{\code{cluster}}{Numeric vector indicating the cluster groups for each of the \eqn{p} original variables}
\item{\code{K}}{Number of aggregated groups detected by tag-lasso}
}
\description{
This function computes the tag-lasso estimator for fixed tuning parameters lambda1 and lambda2
}
